G4TransportationWithMsc: Always update momentum direction (#76)
This fixes a very rare bug when G4Transportation runs without field,
G4TransportationWithMsc is allowed to make multiple internal steps,
and the last internal step does *not* sample scattering (because the
track is either ranging out, or the step is very small). In this case
G4Transportation will reset fMomentumChanged = false because linear
propagation does not change the momentum direction, when in fact
earlier internal steps updated the direction due to scattering.

Instead always update the momentum direction because the above is
rare. Numerical differences are observed in the simulation histories,
but not statistically significant.